# docs/swagger.yaml
definitions:
  models.AllYoutube:
    properties:
      channelID:
        type: string
      channelTitle:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      publishedAt:
        type: string
      query:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      videoID:
        type: string
    type: object
  models.Blog:
    properties:
      bloggerlink:
        type: string
      bloggername:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      postdate:
        type: string
      priority:
        type: integer
      query:
        type: string
      title:
        type: string
    type: object
...
  models.Graph:
    properties:
      areaName:
        type: string
      gradeCode:
        type: string
      graphLine:
        items:
          $ref: '#/definitions/models.GraphLine'
        type: array
      itemCode:
        type: integer
      itemKindCode:
        type: string
      lastShipDate:
        type: string
      rangeLabel:
        items:
          type: string
        type: array
      rangeMax:
        type: integer
      rangeMin:
        type: integer
      rangeStep:
        type: integer
      rangeType:
        type: string
    type: object
  models.GraphData:
    properties:
      x:
        type: string
      "y":
        type: number
    type: object
  models.GraphLine:
    properties:
      areaName:
        type: string
      graphData:
        items:
          $ref: '#/definitions/models.GraphData'
        type: array
      marketName:
        type: string
    type: object
  models.Item:
    properties:
      itemCode:
        type: integer
      itemName:
        type: string
    type: object
  models.ItemGrade:
    properties:
      gradeCode:
        type: string
      gradeName:
        type: string
      gradeOrder:
        type: integer
      gradeType:
        type: integer
      shipAmt:
        type: number
      shipUnit:
        type: string
    type: object
  models.ItemKind:
    properties:
      itemCode:
        type: integer
      itemKindCode:
        type: string
      itemKindName:
        type: string
      itemName:
        type: string
    type: object
...
  models.News:
    properties:
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      press:
        type: string
      priority:
        type: integer
      pubDate:
        type: string
      query:
        type: string
      title:
        type: string
    type: object
  models.RecentPrice:
    properties:
      areaName:
        type: string
      gradeCode:
        type: string
      itemCode:
        type: integer
      itemKindCode:
        type: string
      lastPrice:
        type: integer
      lastShipDate:
        type: string
      marketName:
        type: string
      stat7days:
        $ref: '#/definitions/models.StatPrice'
        type: object
      stat30days:
        $ref: '#/definitions/models.StatPrice'
        type: object
    type: object
  models.StatPrice:
    properties:
      avgPrice:
        type: number
      cntPrice:
        type: integer
      days:
        type: integer
      maxPrice:
        type: integer
      minPrice:
        type: integer
      sumPrice:
        type: integer
    type: object
  models.Youtube:
    properties:
      channelID:
        type: string
      channelTitle:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      publishedAt:
        type: string
      query:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      videoID:
        type: string
    type: object
...
  v2.Blog:
    properties:
      bloggerlink:
        type: string
      bloggername:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      postdate:
        type: string
      priority:
        type: integer
      query:
        type: string
      title:
        type: string
    type: object
...
  v2.News:
    properties:
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      link:
        type: string
      press:
        type: string
      priority:
        type: integer
      pubDate:
        type: string
      query:
        type: string
      title:
        type: string
    type: object
...
  v2.Youtube:
    properties:
      channelID:
        type: string
      channelTitle:
        type: string
      description:
        type: string
      isDisplay:
        type: integer
      itemCode:
        type: string
      publishedAt:
        type: string
      query:
        type: string
      thumbnailURL:
        type: string
      title:
        type: string
      videoID:
        type: string
    type: object
...
info:
  contact: {}
  description: 아그리파 API
  license: {}
  title: AGRIPA API
  version: "1.0"
paths:
  /media/all-youtube:
    get:
      consumes:
      - application/json
      description: 모든 유튜브 영상 조회
      parameters:
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      - description: 검색어(미입력 시 모든 유튜브 검색, 가락시장365만 보고 싶은 경우 '000공통000'입력)
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllYoutube'
      summary: 모든 유튜브 영상 조회
      tags:
      - 미디어
  /media/blog:
    get:
      consumes:
      - application/json
      description: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Blog'
      summary: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
  /media/news:
    get:
      consumes:
      - application/json
      description: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.News'
      summary: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
  /media/youtube:
    get:
      consumes:
      - application/json
      description: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 품목코드
        in: query
        name: itemCode
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: baseDateTime
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: countPerPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Youtube'
      summary: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - 미디어
...
  /retail/grade/{itemcode}/{kindcode}:
    get:
      consumes:
      - application/json
      description: 등급 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemGrade'
      summary: 등급 정보 조회
      tags:
      - 소매
  /retail/grade/{itemcode}/{kindcode}/{gradecode}:
    get:
      consumes:
      - application/json
      description: 지역 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemGrade'
      summary: 지역 정보 조회
      tags:
      - 소매
  /retail/item/{catecode}:
    get:
      consumes:
      - application/json
      description: 품목 정보를 조회합니다
      parameters:
      - description: 부류코드
        in: path
        name: catecode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
      summary: 품목 정보 조회
      tags:
      - 소매
  /retail/kind/{itemcode}:
    get:
      consumes:
      - application/json
      description: 품종 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemKind'
      summary: 품종 정보 조회
      tags:
      - 소매
  /retail/price/graph/{itemcode}/{kindcode}/{gradecode}/{areaname}/{graphType}:
    get:
      consumes:
      - application/json
      description: 소매 가격 정보 - 그래프용
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      - description: 지역명
        in: path
        name: areaname
        required: true
        type: string
      - description: 그래프타입(day,month,year)
        in: path
        name: graphType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Graph'
      summary: 소매 가격 정보(그래프용)
      tags:
      - 소매
  /retail/price/month/{itemcode}/{kindcode}/{gradecode}/{areaname}:
    get:
      consumes:
      - application/json
      description: 소매 월 가격 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      - description: 지역명
        in: path
        name: areaname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecentPrice'
      summary: 소매 월 가격 정보 조회
      tags:
      - 소매
  /retail/price/recent/{itemcode}/{kindcode}/{gradecode}/{areaname}:
    get:
      consumes:
      - application/json
      description: 소매 최근 가격 정보를 조회합니다
      parameters:
      - description: 품목코드
        in: path
        name: itemcode
        required: true
        type: integer
      - description: 품종코드
        in: path
        name: kindcode
        required: true
        type: string
      - description: 등급코드
        in: path
        name: gradecode
        required: true
        type: string
      - description: 지역명
        in: path
        name: areaname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecentPrice'
      summary: 소매 최근 가격 정보 조회
      tags:
      - 소매
...
  /v2/media/blog:
    get:
      consumes:
      - application/json
      description: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Blog'
      summary: 품목별 블로그 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
  /v2/media/news:
    get:
      consumes:
      - application/json
      description: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.News'
      summary: 품목별 뉴스 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
  /v2/media/youtube:
    get:
      consumes:
      - application/json
      description: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      parameters:
      - description: 표준품목코드
        in: query
        name: std-item-code
        required: true
        type: integer
      - description: 기준시간(YYYY-MM-DD HH:mm:ss)
        in: query
        name: base-date-time
        required: true
        type: string
      - description: '조회 시작 index(허용값: 최소 0)'
        in: query
        minimum: 0
        name: start
        type: integer
      - description: '페이지 당 자료 수(허용값: 최소 0, 최대 50)'
        in: query
        maximum: 50
        minimum: 0
        name: count-per-page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Youtube'
      summary: 품목별 유튜브 조회(기준시간으로부터 6개월까지만 조회 가능)
      tags:
      - v2 미디어
...
swagger: "2.0"